- fixed crashes - extracted string resources - updated file directories to fit a standard that i swear exists but i forgot it
