+++ app/src/main/AndroidManifest.xml
@@ -13,14 +13,14 @@
         tools:ignore="GoogleAppIndexingWarning"
         android:fullBackupContent="false">
         <activity
-            android:name=".activities.SplashScreenActivity"
+            android:name=".ui.SplashScreenActivity"
             android:theme="@style/SplashTheme">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".activities.MainActivity">
+        <activity android:name=".ui.main.MainActivity">
         </activity>
     </application>
